cpu: rv64: consolidate JIT kernel files
