+++ src/cpu/rv64/jit_rvv_gemm_convolution_kernel.hpp
@@ -14,8 +14,8 @@
 * limitations under the License.
 *******************************************************************************/
 
-#ifndef CPU_RV64_JIT_RVV_GEMM_CONVOLUTION_POST_KERNEL_HPP
-#define CPU_RV64_JIT_RVV_GEMM_CONVOLUTION_POST_KERNEL_HPP
+#ifndef CPU_RV64_JIT_RVV_GEMM_CONVOLUTION_KERNEL_HPP
+#define CPU_RV64_JIT_RVV_GEMM_CONVOLUTION_KERNEL_HPP
 
 #include "common/c_types_map.hpp"
 #include "cpu/rv64/jit_generator.hpp"
@@ -24,6 +24,27 @@ namespace dnnl {
 namespace impl {
 namespace cpu {
 namespace rv64 {
+
+struct jit_rvv_gemm_convolution_copy_kernel_t : public jit_generator_t {
+    struct call_params_t {
+        const float *src;
+        float *dst;
+        dim_t len;
+    };
+
+    DECLARE_CPU_JIT_AUX_FUNCTIONS(jit_rvv_gemm_convolution_copy_kernel_t)
+
+    jit_rvv_gemm_convolution_copy_kernel_t();
+
+    void operator()(const call_params_t *p) const {
+        jit_generator_t::operator()(p);
+    }
+
+protected:
+    void generate() override;
+};
+
+void jit_rvv_gemm_convolution_copy_f32(const float *src, float *dst, dim_t len);
 
 struct jit_rvv_gemm_convolution_post_kernel_t : public jit_generator_t {
     struct call_params_t {
@@ -68,4 +89,4 @@ void jit_rvv_gemm_convolution_apply_scalar_bias_relu(
 } // namespace impl
 } // namespace dnnl
 
-#endif // CPU_RV64_JIT_RVV_GEMM_CONVOLUTION_POST_KERNEL_HPP
+#endif // CPU_RV64_JIT_RVV_GEMM_CONVOLUTION_KERNEL_HPP
